Now hippie gets loaded from a .onnx
(I really hope this works because I spended 2 days and I need it to be working for the thesis or if not ima delete this whole repository)
Streamlit couldn't use some packages since they were "private" to the lab. Now it uses onnx to load the hippie model, since streamlit can now install onnx library in the environment for the web. This also boosts performance since loading an onnx file is way faster :).